Make StreetName component dumber by moving functionality to StreetNameCanvas
